Last merge fixes
Fixed PI and added comments

implicit CHECKP in getTextPointer()

System exception on wrong CHECKP argument in debug mode

getFirstDescandantBySchema fixed and comments added

getSchemaNode always causes CHECKP now: anyway it calls catalog object search in hashtable

same_node_fair: brain overheating bug fixed

get_base_uri_deep:checkp added

SQL connection fixed and some warnings fixed

Text insertion fixed;
Some warnings fixed;
Datastructure version variable incremented;

getActualParentNode

base-uri fix

preceding axis fixed; dm:base-uri fixes

Base-uri fix

+++ kernel/tr/idx/btree/btstruct.h
@@ -146,7 +146,7 @@ public:
     void * data () const { return ((type == xs_string )?(void*)(v.s_v):(void*)&v); }
 
     xmlscm_type get_type() const { return type; }
-    int get_size() const;
+    size_t get_size() const;
 };
